pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.version]
path = "main.py"
[tool.hatch.build.targets.wheel]
packages = ["src"]
[project]
name = "project_visualization_tool_GDeLuisi"
version = "__VERSION__"
authors = [
{ name="De Luisi Gerardo", email="deluisigerardo@gmail.com" },
]
description = "Visualization tool for git-managed projects"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
"Programming Language :: Python :: 3.10",
"Programming Language :: Python :: 3.11",
"Programming Language :: Python :: 3.12",
"Programming Language :: Python :: 3.13",
"License :: OSI Approved :: MIT License",
"Operating System :: OS Independent",
]
keywords = ["git","projects","visualization-tool"]
dependencies = [
    "blinker",
    "build",
    "certifi",
    "charset-normalizer",
    "click",
    "colorama",
    "coverage",
    "dash",
    "dash-bootstrap-components",
    "dash-core-components",
    "dash-html-components",
    "dash-table",
    "dill",
    "flask",
    "gitdb",
    "gitpython",
    "idna",
    "importlib-metadata",
    "iniconfig",
    "itsdangerous",
    "jinja2",
    "lizard",
    "markupsafe",
    "multiprocess",
    "nest-asyncio",
    "numpy",
    "packaging",
    "pandas",
    "plotly",
    "pluggy",
    "psutil",
    "pydriller",
    "pygments",
    "pyproject-hooks",
    "python-dateutil",
    "pytz",
    "requests",
    "retrying",
    "setuptools",
    "six",
    "smmap",
    "tenacity",
    "types-pytz",
    "typing-extensions",
    "tzdata",
    "urllib3",
    "waitress",
    "werkzeug",
    "zipp"
]
[tool.pytest.ini_options]
pythonpath = [
  ".", "src","src/_internal"
]
testpaths = [
  "tests"
]
[tool.coverage.run]
omit = [
  "src/utility/logs.py",
  "main.py",
  "tests.py",
  "src/gui/*",
  "src/app/app.py"
]
[project.optional-dependencies]
# development dependency groups
test = [
    'pytest',
    'pytest-cov',
]
[project.scripts]
project-viewer = "src.app.cli:main"
[project.urls]
Homepage = "https://github.com/GDeLuisi/project_visualization_tool"
Issues = "https://github.com/GDeLuisi/project_visualization_tool/issues"
Repository = "https://github.com/GDeLuisi/project_visualization_tool"
